Potential fix to issue #1
The nav pulldown is now very good at being resolution-agnostic. This comes with one annoying caveat: the behavior when scrolling down past the nav pulldown while it's open may be confusing to some users. Also, the open-nav hamburger button now moves the focus to the top of the screen. This behavior can't be avoided if the nav is allowed to be scrolled.

Tomorrow I plan to try another potential fix (using a <select> tag for small-resolution navigation) and see which I prefer.